backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
boto3==1.37.10
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
